Moved the voterGuide sub-files into the components tree.
