+++ src/js/components/VoterGuide/VoterGuideFollowers.jsx
@@ -1,8 +1,8 @@
 import React, {Component, PropTypes } from "react";
 import Helmet from "react-helmet";
 import GuideStore from "../../stores/GuideStore";
-import GuideList from "../../components/VoterGuide/GuideList";
-import LoadingWheel from "../../components/LoadingWheel";
+import GuideList from "./GuideList";
+import LoadingWheel from "../LoadingWheel";
 import VoterStore from "../../stores/VoterStore";
 var _ = require("lodash");
